Use shelve instead of dbhash
dbhash is not loadable on Mac OS X because of a missing BSD database
library (bsddb.)

Fortunately, shelve works, and since this is just a dictionary that is
persisted to disk, it's perfect for this use case.

Note: shelve already appends '.db' to the end of the filename, so remove
it from our script.